Resource Reward Tests Complete
Only engagement cycle testing left. I changed up the Depleted pool rewarding and split it into a Resource Tranfer to Pool function that checks both the depleted and resource pool values at the same time.

+++ WarGame/Assets/Scripts/GUI/PoolContribute.cs
@@ -67,8 +67,8 @@ public class PoolContribute: WorldGUI
             }
 
             modifiers[poolType - 1] = 0;
-            modifiers[tempTypeA] = (double)tempPools[tempTypeB] / tempPools[tempTypeA];
-            modifiers[tempTypeB] = (double)tempPools[tempTypeA] / tempPools[tempTypeB];
+            modifiers[tempTypeA] = tempPools[tempTypeB] / tempPools[tempTypeA];
+            modifiers[tempTypeB] = tempPools[tempTypeA] / tempPools[tempTypeB];
         }
 
         strModifiers[0] = string.Format("x {0:0.000}", modifiers[0]);
